!feat: overhaul weight loading and recording
as it is now it seeeeeems to be working for multiprocessing as well (since the from the Connect call onwards it's NEST's problem (which VP the synapse is actually in))
i'll need a bit more time to verify whether weights are actually moving.
we need some place for parameters to live in (i think population_params is suitable), and !!!!!!!!!!! this relies on changing bsb_nest.adapter.py:176:
+ if (connection_model.get_syn_spec()["synapse_model"] == "static_synapse"):

of course, this is rather brittle. we can speak on how and if to improve this once we know the approach is OK

lastly, it may be that with these changes the simulation has slowed down a bit; this surprised be, because i didn't really touch anything on the simulation side, just data loading and saving. will do a complete check at the end.